Change to babel-eslint
For example class methods implemented as arrow functions are not part of
the javascript standard. It's an experimental feature in eslint but they
can be allowed by using babel-eslint.

React says that it's one way around binding functions in classes noting
that it's experimental: https://reactjs.org/docs/handling-events.html